Separated out the java-net rules to create direct rules as independent modules that can be 'plugged in' in the same way as the rules for integration frameworks (eventually)
