+++ opentracing-agent-itests/src/main/java/io/opentracing/contrib/agent/common/OTAgentTestBase.java
