+++ rules/java-net/pom.xml
@@ -2,12 +2,12 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <artifactId>opentracing-agent-parent</artifactId>
+    <artifactId>opentracing-agent-rules-parent</artifactId>
     <groupId>io.opentracing.contrib</groupId>
     <version>0.0.1-SNAPSHOT</version>
   </parent>
 
-  <artifactId>opentracing-agent-itests</artifactId>
+  <artifactId>opentracing-agent-rules-java-net</artifactId>
 
   <properties>
     <opentracing-agent.lib>${project.build.directory}/lib</opentracing-agent.lib>
@@ -16,11 +16,8 @@
   <dependencies>
     <dependency>
       <groupId>io.opentracing</groupId>
-      <artifactId>opentracing-mock</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.opentracing.contrib</groupId>
-      <artifactId>opentracing-globaltracer</artifactId>
+      <artifactId>opentracing-api</artifactId>
+      <scope>provided</scope>
     </dependency>
 
     <dependency>
@@ -29,31 +26,17 @@
       <scope>test</scope>
       <version>${project.version}</version>
     </dependency>
+
     <dependency>
       <groupId>io.opentracing.contrib</groupId>
-      <artifactId>opentracing-web-servlet-filter</artifactId>
+      <artifactId>opentracing-agent-itests</artifactId>
       <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.opentracing.contrib</groupId>
-      <artifactId>opentracing-okhttp3</artifactId>
-      <scope>test</scope>
+      <version>${project.version}</version>
     </dependency>
 
     <dependency>
-      <groupId>com.squareup.okhttp3</groupId>
-      <artifactId>okhttp</artifactId>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-server</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-servlet</artifactId>
+      <groupId>io.opentracing</groupId>
+      <artifactId>opentracing-mock</artifactId>
       <scope>test</scope>
     </dependency>
 
@@ -67,6 +50,7 @@
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>${version.junit}</version>
+      <scope>test</scope>
     </dependency>
   </dependencies>
