+++ rules/java-net/src/main/java/io/opentracing/contrib/agent/jnet/propagation/HttpURLConnectionInjectAdapter.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2015-2017 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.opentracing.contrib.agent.jnet.propagation;
+
+import java.net.HttpURLConnection;
+import java.util.Iterator;
+import java.util.Map;
+
+import io.opentracing.propagation.TextMap;
+
+/**
+ * Inject adapter for use with {@link HttpURLConnection}.
+ */
+public final class HttpURLConnectionInjectAdapter implements TextMap {
+
+    private final HttpURLConnection connection;
+
+    public HttpURLConnectionInjectAdapter(Object connection) {
+        this.connection = (HttpURLConnection)connection;
+    }
+
+    @Override
+    public Iterator<Map.Entry<String, String>> iterator() {
+        throw new UnsupportedOperationException("HttpURLConnectionInjectAdapter should only be used with Tracer.inject()");
+    }
+
+    @Override
+    public void put(String key, String value) {
+        connection.addRequestProperty(key, value);
+    }
+}
